Create SQL Project 1
Understanding SQL is a project that I am embarking on in between studying for security+ and what I will study next semester. Database is a very important subject to understand if I have to become a database administrator. 
Database is a container that helps organize data and SQL statements are the syntaxes that are used to answer all the questions that are queried. 
In this project I created a database and INSERTED Values into the contacts table and answered all the questions (Queries) and at the end, I deleted the table.